Fix doctor service startup and configuration

+++ Backend-Server/WeCureIT/Doctor/src/main/java/com/wecureit/doctor/model/response/ChamberResponse.java
@@ -1,9 +1,13 @@
 package com.wecureit.doctor.model.response;
 
 import lombok.*;
+import java.math.BigDecimal;
 import java.util.List;
 
-@Getter @Setter @NoArgsConstructor @AllArgsConstructor
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class ChamberResponse {
     private Long id;
     private Long facilityId;
@@ -13,6 +17,6 @@ public class ChamberResponse {
     private String slotStart;
     private String slotEnd;
     private Integer slotDurationMinutes;
-    private Double consultationCharge;
+    private BigDecimal consultationCharge;
     private List<String> availableDays;
 }
